pyproject.toml
[tool.poetry]
name = "addresses-heat-maps"
version = "0.0.1"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "addresses_heat_maps"}]

[tool.poetry.dependencies]
python = "^3.10"
frozendict = "^2.3.8"
sqlalchemy = "^2.0.17"
pyyaml = "^6.0"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
